Provide a gevent wrapper for aio_read and aio_write
changed the pyaio.c to core.c so that we can also include pure python
modules in the same package space.

I removed as much memory copying as I could by implementing the new
python buffer protocol and using memoryview objects. Any buffer protocol
supporting object can be a write buffer now.

I added an if to setup.py to reject python less than 2.7

I added exception raising to the aio_read and aio_write dispatch
methods, so we know what kind of errors fired instead of queuing the aio

added gevent.py which contains the class aioFile it is a file like
object that uses gevent and aio_read and aio_write

I updated the tests, used Semaphores to determine when the events are
done.